refactor: adopt schema-owned SwiftData model aliases
Move the V1 SwiftData models into BaseChatSchemaV1, preserve the public top-level names as typealiases, and update previews/examples/docs to prefer ModelContainerFactory. Add regression coverage for alias identity, alias/schema interchangeability, and APIEndpoint persistence.
