+++ Example/Examples/MinimalExample/README.md
@@ -5,7 +5,7 @@ The simplest possible BaseChatKit app. Demonstrates the bare minimum setup:
 - Configure `BaseChatConfiguration` at startup
 - Register backends with `DefaultBackends.register(with:)`
 - Create a `ChatViewModel` and `SessionManagerViewModel`
-- Set up SwiftData with `BaseChatSchema.allModelTypes`
+- Set up SwiftData with `ModelContainerFactory.makeInMemoryContainer()`
 - Present `ChatView` with environment wiring
 
 ## Running
