+++ Sources/BaseChatCore/BaseChatSchemaV1.swift
@@ -1,3 +1,4 @@
+import Foundation
 import SwiftData
 
 /// Version 1 of the BaseChatKit SwiftData schema.
@@ -30,6 +31,218 @@ public enum BaseChatSchemaV1: VersionedSchema {
             SamplerPreset.self,
             APIEndpoint.self,
         ]
+    }
+
+    /// A single message in a chat conversation, persisted via SwiftData.
+    ///
+    /// Messages belong to a session (identified by `sessionID`).
+    @Model
+    public final class ChatMessage {
+        public var id: UUID
+        public var role: MessageRole
+        public var content: String
+        public var timestamp: Date
+        public var sessionID: UUID
+
+        /// Tokens used in the prompt for this response (cloud API backends only).
+        public var promptTokens: Int?
+        /// Tokens generated in this response (cloud API backends only).
+        public var completionTokens: Int?
+
+        public init(
+            role: MessageRole,
+            content: String,
+            sessionID: UUID
+        ) {
+            self.id = UUID()
+            self.role = role
+            self.content = content
+            self.timestamp = Date()
+            self.sessionID = sessionID
+        }
+    }
+
+    /// A chat session containing a sequence of messages with its own settings.
+    ///
+    /// Sessions hold per-session overrides for generation parameters. When an
+    /// override is `nil`, the app falls back to global defaults from `SettingsService`.
+    @Model
+    public final class ChatSession {
+        public var id: UUID
+        public var title: String
+        public var createdAt: Date
+        public var updatedAt: Date
+
+        /// Per-session system prompt.
+        public var systemPrompt: String
+
+        /// The UUID of the selected ModelInfo for this session.
+        public var selectedModelID: UUID?
+
+        /// The UUID of the selected APIEndpoint for this session.
+        public var selectedEndpointID: UUID?
+
+        // Per-session generation overrides (nil = use global default)
+        public var temperature: Float?
+        public var topP: Float?
+        public var repeatPenalty: Float?
+
+        /// Stored as PromptTemplate.rawValue; nil means auto-detect or global default.
+        public var promptTemplateRawValue: String?
+
+        /// User override for context window size; nil uses model default.
+        public var contextSizeOverride: Int?
+
+        /// Raw storage for CompressionMode. nil means .automatic.
+        /// SwiftData lightweight migration handles this new optional column automatically.
+        public var compressionModeRaw: String?
+
+        /// Comma-separated UUID strings of pinned messages in this session.
+        /// nil means no messages are pinned.
+        public var pinnedMessageIDsRaw: String?
+
+        public init(title: String = "New Chat") {
+            self.id = UUID()
+            self.title = title
+            self.createdAt = Date()
+            self.updatedAt = Date()
+            self.systemPrompt = ""
+        }
+
+        /// The set of pinned message IDs for this session.
+        ///
+        /// Pinned messages are preserved during context compression regardless of age.
+        /// Serialized as comma-separated UUID strings in ``pinnedMessageIDsRaw``.
+        public var pinnedMessageIDs: Set<UUID> {
+            get {
+                guard let raw = pinnedMessageIDsRaw else { return [] }
+                return Set(raw.split(separator: ",").compactMap { UUID(uuidString: String($0)) })
+            }
+            set {
+                pinnedMessageIDsRaw = newValue.isEmpty ? nil : newValue.map(\.uuidString).joined(separator: ",")
+            }
+        }
+
+        /// The compression mode for this session.
+        ///
+        /// Defaults to `.automatic` when no value is stored.
+        public var compressionMode: CompressionMode {
+            get { compressionModeRaw.flatMap(CompressionMode.init(rawValue:)) ?? .automatic }
+            set { compressionModeRaw = newValue.rawValue }
+        }
+
+        /// Convenience to get/set the prompt template as a `PromptTemplate` enum.
+        public var promptTemplate: PromptTemplate? {
+            get {
+                guard let raw = promptTemplateRawValue else { return nil }
+                return PromptTemplate(rawValue: raw)
+            }
+            set {
+                promptTemplateRawValue = newValue?.rawValue
+            }
+        }
+    }
+
+    /// A saved set of generation parameters that can be applied to any session.
+    @Model
+    public final class SamplerPreset {
+        public var id: UUID
+        public var name: String
+        public var temperature: Float
+        public var topP: Float
+        public var repeatPenalty: Float
+        public var createdAt: Date
+
+        public init(name: String, temperature: Float = 0.7, topP: Float = 0.9, repeatPenalty: Float = 1.1) {
+            self.id = UUID()
+            self.name = name
+            self.temperature = temperature
+            self.topP = topP
+            self.repeatPenalty = repeatPenalty
+            self.createdAt = Date()
+        }
+    }
+
+    /// A configured cloud API endpoint persisted via SwiftData.
+    ///
+    /// The API key is NOT stored here — it lives in the Keychain, referenced
+    /// by this endpoint's `id` as the Keychain account identifier.
+    @Model
+    public final class APIEndpoint {
+        public var id: UUID
+        public var name: String
+        public var providerRawValue: String
+        public var baseURL: String
+        public var modelName: String
+        public var createdAt: Date
+        public var isEnabled: Bool
+
+        public init(
+            name: String,
+            provider: APIProvider,
+            baseURL: String? = nil,
+            modelName: String? = nil
+        ) {
+            self.id = UUID()
+            self.name = name
+            self.providerRawValue = provider.rawValue
+            self.baseURL = baseURL ?? provider.defaultBaseURL
+            self.modelName = modelName ?? provider.defaultModelName
+            self.createdAt = Date()
+            self.isEnabled = true
+        }
+
+        /// The provider type as an enum.
+        public var provider: APIProvider {
+            get { APIProvider(rawValue: providerRawValue) ?? .custom }
+            set { providerRawValue = newValue.rawValue }
+        }
+
+        /// The Keychain account identifier for this endpoint's API key.
+        public var keychainAccount: String {
+            id.uuidString
+        }
+
+        /// Retrieves the API key from the Keychain.
+        public var apiKey: String? {
+            KeychainService.retrieve(account: keychainAccount)
+        }
+
+        /// Stores an API key in the Keychain for this endpoint.
+        @discardableResult
+        public func setAPIKey(_ key: String) -> Bool {
+            KeychainService.store(key: key, account: keychainAccount)
+        }
+
+        /// Deletes the API key from the Keychain.
+        public func deleteAPIKey() {
+            KeychainService.delete(account: keychainAccount)
+        }
+
+        /// Validates the endpoint configuration.
+        public var isValid: Bool {
+            guard let url = URL(string: baseURL), url.scheme != nil, url.host != nil else {
+                return false
+            }
+
+            // HTTPS required for remote endpoints
+            if !isLocalhost(url) && url.scheme != "https" {
+                return false
+            }
+
+            // API key required for providers that need one
+            if provider.requiresAPIKey && (apiKey?.isEmpty ?? true) {
+                return false
+            }
+
+            return true
+        }
+
+        /// Checks if the URL points to a local server.
+        private func isLocalhost(_ url: URL) -> Bool {
+            guard let host = url.host() else { return false }
+            return host == "localhost" || host == "127.0.0.1" || host == "::1" || host == "[::1]"
+        }
     }
 }
