+++ TESTING.md
@@ -535,9 +535,7 @@ final class MyFeatureTests: XCTestCase {
 
     override func setUp() async throws {
         try await super.setUp()
-        let schema = Schema(BaseChatSchema.allModelTypes)
-        let config = ModelConfiguration(isStoredInMemoryOnly: true)
-        container = try ModelContainer(for: schema, configurations: [config])
+        container = try makeInMemoryContainer()
         context = container.mainContext
 
         let backend = MockInferenceBackend()
